// src/Enums/EBoxen.enum.ts
// Enumération des noms de couleurs disponibles
export enum ColorName {
    BLACK = 'black',
    RED = 'red',
    GREEN = 'green',
    YELLOW = 'yellow',
    BLUE = 'blue',
    MAGENTA = 'magenta',
    CYAN = 'cyan',
    WHITE = 'white',
    GREY = 'grey',
}

// Enumération des styles de bordure disponibles
export enum BorderStyleName {
    SINGLE = 'single',
    DOUBLE = 'double',
    ROUND = 'round',
    BOLD = 'bold',
    SINGLE_DOUBLE = 'singleDouble',
    DOUBLE_SINGLE = 'doubleSingle',
    CLASSIC = 'classic',
}

// Enumération des alignements disponibles
export enum AlignementName {
    LEFT = 'left',
    CENTER = 'center',
    RIGHT = 'right',
}

// Enumération des messages de titre disponibles
export enum TitleMessage {
    INFORMATION = 'IPOView - Information',
    WARNING = 'IPOView - Warning',
    ERROR = 'IPOView - Error',
    SUCCESS = 'IPOView - Success',
    DIVERS = 'IPOView - Divers',
    BIENVENUE = 'IPOView - Bienvenue',
}